requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==0.8.11
requests==2.31.0
beautifulsoup4==4.12.2
ollama==0.1.7
python-dotenv==1.0.0

